rename labelize into label
